Reorganized Python project. Changed build system. Fixed module imports. Added dynamic version
