+++ .gitignore
@@ -172,6 +172,9 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
+# VSCode
+.vscode
+
 # Ruff stuff:
 .ruff_cache/
