--- odoomap.py
@@ -1,5 +1,0 @@
-# Just a wrapper for core.py
-from src.core import main
-
-if __name__ == "__main__":
-    main()
