+++ odoomap/__main__.py
@@ -1,0 +1,4 @@
+# Just a wrapper for core.py
+from .core import main
+
+main()
