+++ odoomap/connect.py
@@ -5,9 +5,10 @@ import ssl
 import urllib3
 import os
 from bs4 import BeautifulSoup
-from src.colors import Colors
 from urllib.parse import urljoin
 from importlib.resources import files
+
+from .colors import Colors
 
 # Disable SSL warnings
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
@@ -100,10 +101,10 @@ class Connection:
         user_pass_pairs = []
 
         try:
-            usernames_text = files("src.data").joinpath("default_usernames.txt").read_text(encoding='utf-8', errors='ignore')
+            usernames_text = files("odoomap.data").joinpath("default_usernames.txt").read_text(encoding='utf-8', errors='ignore')
             usernames = [line.strip() for line in usernames_text.splitlines() if line.strip()]
 
-            passwords_text = files("src.data").joinpath("default_passwords.txt").read_text(encoding='utf-8', errors='ignore')
+            passwords_text = files("odoomap.data").joinpath("default_passwords.txt").read_text(encoding='utf-8', errors='ignore')
             passwords = [line.strip() for line in passwords_text.splitlines() if line.strip()]
 
         except Exception as e:
